requirements.txt
numpy
pandas
setuptools
scikit-learn
lightgbm
matplotlib
mlflow
flask
pylint
flake8
black
tensorboard
torch
matplotlib
my-sql-connector-python
xgboost
mlflow
dvc
-e .